+++ api/src/queries/activities-v2-queries.ts
@@ -1,84 +1,11 @@
-import { getuid } from 'process';
-import { Operation } from 'express-openapi';
-import { RequestHandler } from 'express';
 import SQL, { SQLStatement } from 'sql-template-strings';
 import { escapeLiteral } from 'pg';
 import { validActivitySortColumns } from 'sharedAPI/src/misc/sortColumns';
 import { getLogger } from 'utils/logger';
-import { streamActivitiesResult } from 'utils/iapp-json-utils';
-import { getDBConnection } from 'database/db';
-import { ALL_ROLES, SECURITY_ON } from 'constants/misc';
-import { InvasivesRequest } from 'utils/auth-utils';
 
-const defaultLog = getLogger('activity');
+const defaultLog = getLogger('activities-v2-queries');
 
-export const POST: Operation = [getActivitiesBySearchFilterCriteria()];
-
-POST.apiDoc = {
-  description: 'Fetches all activities based on search criteria.',
-  tags: ['activity'],
-  security: SECURITY_ON
-    ? [
-        {
-          Bearer: ALL_ROLES
-        }
-      ]
-    : [],
-  requestBody: {
-    description: 'Activities Request Object',
-    content: {
-      'application/json': {
-        schema: {
-          properties: {}
-        }
-      }
-    }
-  },
-  responses: {
-    200: {
-      description: 'Activity get response object array.',
-      content: {
-        'application/json': {
-          schema: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                rows: {
-                  type: 'array',
-                  items: {
-                    type: 'object',
-                    properties: {
-                      // Don't specify exact object properties, as it will vary, and is not currently enforced anyways
-                      // Eventually this could be updated to be a oneOf list, similar to the Post request below.
-                    }
-                  }
-                },
-                count: {
-                  type: 'number'
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    304: {
-      $ref: '#/components/responses/304'
-    },
-    401: {
-      $ref: '#/components/responses/401'
-    },
-    503: {
-      $ref: '#/components/responses/503'
-    },
-    default: {
-      $ref: '#/components/responses/default'
-    }
-  }
-};
-
-export function sanitizeActivityFilterObject(filterObject: any, req: any) {
+function sanitizeActivityFilterObject(filterObject: any, req: any) {
   const sanitizedSearchCriteria = {
     serverSideNamedFilters: {},
     selectColumns: [],
@@ -312,6 +239,10 @@ export function sanitizeActivityFilterObject(filterObject: any, req: any) {
   //todo actually validate:
   sanitizedSearchCriteria.isCSV = filterObject?.isCSV;
   sanitizedSearchCriteria.CSVType = filterObject?.CSVType;
+
+  // may be explicitly set to true by calling function
+  sanitizedSearchCriteria.boundingBoxOnly = false;
+
   defaultLog.debug({
     label: 'getActivitiesBySearchFilterCriteria',
     message: 'sanitizedObject',
@@ -321,68 +252,7 @@ export function sanitizeActivityFilterObject(filterObject: any, req: any) {
   return sanitizedSearchCriteria;
 }
 
-/**
- * Fetches all activity records based on request search filter criteria.
- *
- * @return {RequestHandler}
- */
-function getActivitiesBySearchFilterCriteria(): RequestHandler {
-  const reqID = getuid();
-  return async (req: InvasivesRequest, res) => {
-    if (req.authContext.roles.length === 0) {
-      res.status(401).json({ message: 'No Role for user' });
-    }
-
-    const rawBodyCriteria = req.body['filterObjects'];
-    const filterObject = sanitizeActivityFilterObject(rawBodyCriteria?.[0], req);
-    defaultLog.debug({ label: 'v2/activity', message: 'getActivitiesBySearchFilterCriteria v2', body: '' });
-
-    let connection;
-    let sql;
-
-    try {
-      connection = await getDBConnection();
-      if (!connection) {
-        defaultLog.error({
-          label: 'v2/activity',
-          message: 'getActivitiesBySearchFilterCriteria',
-          body: 'reqID:' + reqID + ' - ' + 'Database connection unavailable'
-        });
-        return res.status(503).json({ message: 'Database connection unavailable', namespace: 'activities', code: 503 });
-      }
-
-      sql = getActivitiesSQLv2(filterObject);
-
-      if (filterObject.isCSV && filterObject.CSVType) {
-        res.status(200);
-        await streamActivitiesResult(filterObject, res, sql);
-      } else {
-        const response = await connection.query(sql.text, sql.values);
-
-        return res.status(200).json({
-          message: 'fetched activities by criteria',
-          request: req.body,
-          result: response.rows,
-          count: response.rowCount,
-          namespace: 'activities',
-          code: 200
-        });
-      }
-    } catch (error) {
-      defaultLog.debug({ label: 'getActivitiesBySearchFilterCriteria', message: 'error', error });
-      return res.status(500).json({
-        message: 'Error getting activities by search filter criteria',
-        error,
-        namespace: 'v2/activities',
-        code: 500
-      });
-    } finally {
-      connection?.release();
-    }
-  };
-}
-
-export function getActivitiesSQLv2(filterObject: any) {
+function getActivitiesSQLv2(filterObject: any) {
   defaultLog.debug({
     label: 'getActivitiesBySearchFilterCriteria',
     message: 'sql',
@@ -396,12 +266,14 @@ export function getActivitiesSQLv2(filterObject: any) {
     sqlStatement = fromStatement(sqlStatement, filterObject);
     sqlStatement = whereStatement(sqlStatement, filterObject);
     sqlStatement = groupByStatement(sqlStatement, filterObject);
-    if (!filterObject.vt_request) {
+    if (filterObject.vt_request) {
+      sqlStatement.append(` ) SELECT ST_AsMVT(mvtgeom.*, 'data', 4096, 'geom', 'feature_id') as data from mvtgeom;`);
+    } else if (filterObject.boundingBoxOnly) {
+      sqlStatement.append(`) SELECT ST_AsText(ST_Extent(geometry(geog))) as bbox;`);
+    } else {
       sqlStatement = orderByStatement(sqlStatement, filterObject);
       sqlStatement = limitStatement(sqlStatement, filterObject);
       sqlStatement = offSetStatement(sqlStatement, filterObject);
-    } else {
-      sqlStatement.append(` ) SELECT ST_AsMVT(mvtgeom.*, 'data', 4096, 'geom', 'feature_id') as data from mvtgeom;`);
     }
 
     defaultLog.debug({ label: 'getActivitiesBySearchFilterCriteria', message: 'sql', body: sqlStatement });
@@ -912,7 +784,7 @@ function whereStatement(sqlStatement: SQLStatement, filterObject: any) {
     where.append(
       ` AND (
           ${tableAlias}.species_biocontrol_full is null
-          OR 
+          OR
           ${tableAlias}.species_biocontrol_full not in (
             SELECT agent_code_description
             FROM invasivesbc.private_biocontrol_agents
@@ -949,3 +821,5 @@ function offSetStatement(sqlStatement: SQLStatement, filterObject: any) {
   const offset = sqlStatement.append(` offset ${filterObject.offset};`);
   return offset;
 }
+
+export { sanitizeActivityFilterObject, getActivitiesSQLv2 };
